@@ -1,28 +1,19 @@
 package appersonal.development.com.appersonaltrainer.Activities;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.ComponentName;
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.CompoundButton;
-import android.widget.ImageView;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
 import android.widget.Switch;
-import android.widget.Toast;
 
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
 
-import appersonal.development.com.appersonaltrainer.Controller.OnAlarmReceiver;
-import appersonal.development.com.appersonaltrainer.Controller.OnBootReceiver;
 import appersonal.development.com.appersonaltrainer.R;
 
 
