@@ -1,6 +1,6 @@
 package appersonal.development.com.appersonaltrainer.Activities;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import appersonal.development.com.appersonaltrainer.R;
