@@ -2,8 +2,8 @@ package appersonal.development.com.appersonaltrainer.Activities;
 
 import android.annotation.SuppressLint;
 import android.os.Handler;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
