@@ -7,11 +7,10 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.Calendar;
 
-import appersonal.development.com.appersonaltrainer.Activities.AlarmeActivity;
 import appersonal.development.com.appersonaltrainer.Model.Refeicoes;
 
 public class BackgroundService extends Service {
